+++ tests/test-options-parse.cpp
@@ -274,7 +274,7 @@ void test_random_perms()
         argv[args.size()] = nullptr;
         for(std::vector<std::string>::const_iterator arg = args.begin(); arg != args.end(); ++arg)
             argv[arg - args.begin()] = arg->c_str();
-        options_t(args.size(), const_cast<char **>(argv));
+        options_t((int) args.size(), const_cast<char **>(argv));
         delete[] argv;
     }
 }
